requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.1.0
openpyxl>=3.1.2
pyjwt>=2.8.0
firebase-admin>=6.5.0
pyrebase4>=4.8.0
gunicorn>=21.2.0
